Homework2/src/cropper/requirements.txt
pillow
piexif
requests
requests_aws4auth